reformated code to work with google styles
